Code review changes + code formatting
