+++ PersonApi/V1/Factories/ISnsFactory.cs
@@ -1,0 +1,10 @@
+using PersonApi.V1.Domain;
+using PersonApi.V1.Infrastructure.JWT;
+
+namespace PersonApi.V1.Factories
+{
+    public interface ISnsFactory
+    {
+        PersonSns Create(Person person, Token token);
+    }
+}
