+++ PersonApi/V1/Factories/ResponseFactory.cs
@@ -2,41 +2,9 @@ using PersonApi.V1.Boundary;
 using PersonApi.V1.Boundary.Response;
 using PersonApi.V1.Domain;
 using System;
-using PersonApi.V1.Infrastructure.JWT;
 
 namespace PersonApi.V1.Factories
 {
-
-
-    public interface ISnsFactory
-    {
-        PersonSns Create(Person person, Token token);
-    }
-
-    public class PersonSnsFactory : ISnsFactory
-    {
-        public PersonSns Create(Person person, Token token)
-        {
-            return new PersonSns
-            {
-                CorrelationId = Guid.NewGuid(),
-                DateTime = DateTime.UtcNow,
-                EntityId = person.Id,
-                Id = Guid.NewGuid(),
-                EventType = "PersonCreatedEvent",
-                Version = "v1",
-                SourceDomain = "Person",
-                SourceSystem = "PersonAPI",
-                User = new User
-                {
-                    Id = Guid.NewGuid(),
-                    Name = token.Name,
-                    Email = token.Email
-                }
-            };
-        }
-    }
-
     public class ResponseFactory : IResponseFactory
     {
         private readonly IApiLinkGenerator _apiLinkGenerator;
