+++ PersonApi/V1/Infrastructure/Constants.cs
@@ -1,0 +1,11 @@
+namespace PersonApi.V1.Infrastructure
+{
+    public static class Constants
+    {
+        // JWT TOKEN
+        public const string V1_VERSION = "v1";
+        public const string EVENTTYPE = "PersonCreatedEvent";
+        public const string SOURCE_DOMAIN = "Person";
+        public const string SOURCE_SYSTEM = "PersonAPI";
+    }
+}
