+++ PersonApi/appsettings.LocalDevelopment.json
@@ -11,6 +11,6 @@
   },
   "DynamoDb": {
     "LocalMode": true,
-    "LocalServiceUrl": "http://localhost:4566"
+    "LocalServiceUrl": "http://localhost:8000"
   }
 }
